Separate API functions from components.
